reorganize Articles feature into sub-components
