+++ src/components/MarkdownArticle.tsx
@@ -2,10 +2,10 @@ import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { Card, CardContent, Container } from "@mui/material";
 import { DpgMarkdown } from "components/DpgMarkdown";
-import { Article } from "features/Article/ArticleType";
+import { Article } from "features/Articles";
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { api } from "../../services";
+import { api } from "services";
 
 export function MarkdownArticle() {
   const { id } = useParams();
