--- src/features/Article/ArticleType.ts
@@ -1,7 +1,0 @@
-export interface Article {
-  id: number;
-  title: string;
-  updateDate: Date;
-  createDate: string;
-  content: string;
-}
