--- src/features/Article/index.ts
@@ -1,5 +1,0 @@
-import { MarkdownArticle } from "features/Article/MarkdownArticle";
-import { Article } from "features/Article/ArticleType";
-
-export { MarkdownArticle };
-export type { Article };
