+++ src/features/ArticleList/ArticleList.tsx
@@ -9,7 +9,7 @@ import {
   Typography,
 } from "@mui/material";
 import { DpgPageError } from "components/DpgError";
-import { Article } from "features/Article";
+import { Article } from "features/Articles";
 import React, { useEffect, useState } from "react";
 import { Link as RouterLink } from "react-router-dom";
 import { api } from "../../services";
