+++ src/features/Articles/index.ts
@@ -1,0 +1,4 @@
+import { Article, Articles } from "./Articles";
+
+export { Articles };
+export type { Article };
